requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Web automation
playwright==1.40.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI and LLM
openai==1.3.7
google-generativeai==0.3.2

# Data processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
chromadb==0.5.3
sentence-transformers==3.0.1

# Utilities
rich==13.7.0
loguru==0.7.2
python-multipart==0.0.6
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# System monitoring
psutil==5.9.6

# File handling
pathlib2==2.3.7
pypdf==3.17.4
python-docx==1.1.2

# Async support
asyncio-mqtt==0.16.1

# JSON handling
orjson==3.9.10

# HTTP client
httpx==0.25.2

